+++ gsl/multiroots/newton.c
@@ -96,7 +96,7 @@ newton_iterate (void * vstate, gsl_multiroot_function_fdf * fdf, double * x, dou
   
   size_t i;//iteration variable
 
-  size_t n = fdf->n ;
+  int n = fdf->n ;
 
   int nrhs=1;//number of right hand side vectors
